fix: address medium-severity review findings
11. Clear cached_layout on transition cancel (already applied by highs agent)
12. Guard Task ActivityStart with handled_by_task_start flag
13. Move PetState from renderer.rs to pet.rs
14. Fix misleading wall thickness comment
15. Remove unused _theme param from paint_pet_hearts
16. Update 3 stale CLAUDE.md sections
17. Delete lock file after config save
18. Add WallDecor::size() and PlantKind::size() methods
